pyproject.toml
[tool.poetry]
name = "aacr_abstract"
version = "0.1.0"
description = ""
authors = ["a2c_hyogo <hyogo.atsushi.cs@daiichisankyo.co.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0.2"
streamlit = "^1.23.1"
numpy = "^1.24.3"
retrying = "^1.3.4"
openai = "^0.27.8"
pdfminer-six = "^20221105"
pypdf = "^3.9.1"
beautifulsoup4 = "^4.12.2"
selenium = "^4.10.0"
webdriver-manager = "^3.8.6"
tiktoken = "^0.4.0"
gdown = "^4.7.1"
google-api-python-client = "^2.92.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
ipykernel = "^6.23.2"
mypy = "^1.3.0"
pytest = "^7.4.0"
types-requests = "^2.31.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
